+++ src/lib/components/custom/side-bar/nav-favorites.svelte
@@ -19,6 +19,7 @@
 	import { type CloudNote, useCloudNotes } from '$lib/supabase/db/cloudnotes.svelte';
 	import { useCurrentUserWorkspaceContext } from '../user-workspace/userworkspace.svelte';
 	import { PUBLIC_NOTA_FRONTEND_URL } from '$env/static/public';
+	import { ask } from '@tauri-apps/plugin-dialog';
 
 	const sidebar = useSidebar();
 	let showMore = $state(false);
@@ -49,6 +50,15 @@
 	}
 
 	async function trashNote(note: LocalNote | CloudNote) {
+		const shouldDelete = await ask(
+			'This action will put the note in Trash. You can still access it and restore it later.',
+			{
+				title: `Trash ${note.name}`,
+				okLabel: 'Trash',
+				cancelLabel: 'Cancel'
+			}
+		);
+		if (!shouldDelete) return;
 		try {
 			if ('owner' in note) await cloudNotes.moveToTrash(note.id);
 			else await localNotes.trashNote(note);
@@ -58,6 +68,15 @@
 		}
 	}
 	async function deleteNote(note: LocalNote | CloudNote) {
+		const shouldDelete = await ask(
+			'This action will permanently delete the note. Are you sure you want to continue?',
+			{
+				title: `Delete ${note.name}`,
+				okLabel: 'Delete',
+				cancelLabel: 'Cancel'
+			}
+		);
+		if (!shouldDelete) return;
 		try {
 			if ('owner' in note) await cloudNotes.deleteNote(note.id);
 			else await localNotes.deleteNote(note);
